Move Task and subclasses into Task package
